fix: stop rendering any script tags in React trees
next/script still re-renders <script> on locale soft-nav and trips
React 19. Move theme + deferred-font bootstrap into
instrumentation-client (module side effects only).

Verified: Playwright locale/blog navigations → 0 "script tag" console
errors; dark theme still applies after reload.
