+++ instrumentation-client.ts
@@ -1,11 +1,17 @@
 /**
  * Client instrumentation runs before the app becomes interactive.
- * React Grab must load early so hover/⌘C can resolve component stacks.
- * Only active under `next dev` (NODE_ENV === "development").
+ * Keep all side-effect bootstrapping here — never put <script> / next/script
+ * in React trees (locale soft-nav re-renders them and React 19 errors).
  *
  * @see https://nextjs.org/docs/app/api-reference/file-conventions/instrumentation-client
- * @see https://www.react-grab.com
  */
+
+import { applyStoredTheme, promoteDeferredFonts } from "@/shared/theme-init";
+
+// Sync: runs as soon as this module evaluates (before app interactive).
+applyStoredTheme();
+promoteDeferredFonts();
+
 if (process.env.NODE_ENV === "development") {
   import("react-grab")
     .then(() => {
